backend/pyproject.toml
[project]
name = "adk-rag-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"

dependencies = [
    "google-cloud-aiplatform==1.92.0",
    "google-cloud-storage==2.19.0",
    "google-genai==1.14.0",
    "gitpython==3.1.40",
    "google-adk==0.5.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]==0.34.0",
    "python-multipart==0.0.9",
    "deprecated==1.2.18",
    "pypdf2>=3.0.1",
    "python-docx>=1.1.2",
    "requests==2.31.0" 
]


[tool.uv.workspace]
members = [
    "frontend",
]
